fix(sandcastle): work around 2017 VSIXInstaller issue
Side-by-side changes for Visual Studio 2017 break quiet mode installs for
VSIXInstaller. To work around this, we've taken Chocolatey's `Install-Vsix`
function and created an extended version that accepts version and SKU
params in order to identify the setup engine VSIXInstaller needs to use
when more than one version or SKU of Visual Studio is installed on the
system.