fix(retroarch): Clean up during uninstall
When installed to a custom location, the Chocolatey auto-uninstaller
doesn't clean up the install location nor removes the shims and shortcut
created. We need to explicitly tell it to uninstall the zip package and
then remove the shims and shortcut.

Since Chocolatey doesn't (yet) provide package params to the uninstaller,
we'll just always uninstall the shims and delete the shortcut. See
https://github.com/chocolatey/choco/issues/1479