fix(retroarch): support new upstream packaging changes
Since RetroArch.7z now contains a directory with the platform-specific
build inside that, stop installing the app to a "retroarch" directory
and instead install it to the base tools directory unless an alternate
directory is specified as a package parameter. Update paths for stubs
and desktop icon as well, so that they'll point to the appropriate
executables and so they will be properly removed on uninstall.